Enables working login form.
Currently not working fully correctily in my dev environment. The mutation succeeds, but the cookies are not set due to cross-site limitations the way my dev environment is set up. There are a few ways to fix it. It would in theory work if I had https working on my dev environment, but that requires self-signing certificates and that adds extra work on potential contributors.