Removes option to get the List from a ListItem
It didn't make sense and was giving me a headache when it came around to testing what would happen if something went wrong during that. If it was possible to get ListItem objects directly it would make more sense, but currently to get ListItems you have to go through the list.